requirements.txt
Flask
flask-cors
huggingface-hub
opencv-python
torchvision
pillow
torch
numpy